feat: extract Native legacy module as well
